Add weekly availability input component
