+++ client/src/components/User/WeeklyAvailability.jsx
@@ -8,50 +8,29 @@ import {
   deleteTimeObj,
   removeExtraTimeObj,
 } from "../../redux/slicers/availabilitySlice";
-import TimeDropdown from "../Elements/dropdown/TimeDropdown";
 import { useParams } from "react-router";
 import DowDropdown from "../Elements/dropdown/DowDropdown";
+import WeeklyAvailabilityInput from "../Elements/Input/WeeklyAvailabilityInput";
 
 const WeeklyAvailability = () => {
   const availability = useSelector((state) => state.availability.weekly);
   const dispatch = useDispatch();
   const param = useParams();
   const copyIcon = useRef()
-  const [selectedItem, setSelectedItem] = useState("");
-  const [clickedElem, setClickedElem] = useState(null);
   const [clickedCopyElem, setClickedCopyElem] = useState(null);
   const [dowId, setDowId] = useState(null);
-
-  useEffect(() => {
-    //Logic of close time dropdown by click anywhere.
-    if (!clickedElem) return;
-
-    const handleCloseTimeDropdown = (e) => {
-      // console.log({clickedElem});
-      // console.log({e: e.target});
-      if (!(clickedElem === e.target)) {
-        setSelectedItem("");
-        setClickedElem(null);
-      }
-    };
-    document.addEventListener("click", handleCloseTimeDropdown);
-    return () => {
-      document.removeEventListener("click", handleCloseTimeDropdown);
-    };
-  }, [selectedItem, clickedElem]);
   
   useEffect(() => {
     //Logic of close copy dropdown by click anywhere.
     if (!clickedCopyElem) return;
 
     const handleCloseCopyDropdown = (e) => {
-      console.log(copyIcon.current);
-      console.log(e.target);
-      console.log(copyIcon.current.contains(e.target));
+      // console.log(copyIcon.current);
+      // console.log(e.target);
+      // console.log(copyIcon.current.contains(e.target));
       if(!copyIcon.current.contains(e.target)){
         setDowId(null);
       }else{
-
 
       }
     };
@@ -60,12 +39,6 @@ const WeeklyAvailability = () => {
       document.removeEventListener("click", handleCloseCopyDropdown);
     };
   }, [dowId]);
-  
-
-  const displayTimeDropdown = (id, elem) => {
-    setSelectedItem(id);
-    setClickedElem(elem);
-  };
 
   const handleMethod = (e, method, dow, data) => {
     e.preventDefault();
@@ -83,9 +56,10 @@ const WeeklyAvailability = () => {
       setClickedCopyElem(e.currentTarget)
       if (!dowId) {
         setDowId(dow);
-      } else {
-        setDowId(null);
-      }
+      } 
+      // else {
+      //   setDowId(null);
+      // }
     }
   };
 
@@ -139,89 +113,32 @@ const WeeklyAvailability = () => {
                     type="checkbox"
                     checked={eachObj[Object.keys(eachObj)[0]]}
                     onChange={() => dispatch(setCheckBox(Object.keys(eachObj)[0]))}
-                    className=""
                   />
                   {Object.keys(eachObj)[0]}
                 </label>
                 <div className="flex flex-col gap-3">
                   {/* Check each day of week's value (which is boolean). If true, map all elem in time array. */}
                   {availability &&
-                  availability.find(
-                    (elem) => Object.keys(elem)[0] === Object.keys(eachObj)[0]
-                  )[Object.keys(eachObj)[0]] ? (
-                    availability &&
-                    availability
-                      .find(
-                        (elem) =>
-                          Object.keys(elem)[0] === Object.keys(eachObj)[0]
-                      )
-                      .time.map((startEndObj, timeIndex) => {
-                        return (
-                          <div
-                            key={timeIndex}
-                            className="flex items-center gap-3 w-[80%] ml-4"
-                          >
-                            <div className="relative">
-                              <input
-                                type="text"
-                                id={`${
-                                  Object.keys(eachObj)[0]
-                                }+start+${timeIndex}`}
-                                value={
-                                  // If time array is initial state (which means no elem in time array), set the value manually.
-                                  availability && 
-                                    availability.find((elem) => Object.keys(elem)[0] === Object.keys(eachObj)[0]).time[0].start === ""
-                                    ? "09:00"
-                                    : startEndObj.start
-                                }
-                                onClick={(e) => displayTimeDropdown(e.target.id, e.currentTarget)}
-                                onChange={(e) => e.target.value}
-                                className="border border-gray-700 rounded-lg w-16 h-7 text-center focus:border-green-400"
-                              />
-                              {/* If selectedItem state is same as clicked input id, render a TimeDropdown component. */}
-                              {selectedItem === `${Object.keys(eachObj)[0]}+start+${timeIndex}` && (
-                                <TimeDropdown
-                                  position={"start"}
-                                  selectedItem={selectedItem}
-                                  timeIndex={timeIndex}
-                                  from={"weekly"}
-                                />
-                              )}
+                    availability.find((elem) => Object.keys(elem)[0] === Object.keys(eachObj)[0])[Object.keys(eachObj)[0]] 
+                    ? (
+                      availability &&
+                        availability.find((elem) => Object.keys(elem)[0] === Object.keys(eachObj)[0]).time.map((startEndObj, timeIndex) => {
+                          return (
+                            <div key={timeIndex} className="flex items-center gap-3 w-[80%] ml-4" >
+                              <WeeklyAvailabilityInput eachObj={eachObj} timeIndex={timeIndex} availability={availability} startEndObj={startEndObj} position={"start"} />
+                              -
+                              <WeeklyAvailabilityInput eachObj={eachObj} timeIndex={timeIndex} availability={availability} startEndObj={startEndObj} position={"end"} />
+                              <div className="flex justify-center items-center w-7 h-7 hover:bg-gray-200 hover:rounded-full">
+                                <svg onClick={(e) => handleMethod(e, "delete", Object.keys(eachObj)[0], startEndObj)} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-5 h-5">
+                                  <path strokeLinecap="round" strokeLinejoin="round" d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"/>
+                                </svg>
+                              </div>
                             </div>
-                            -
-                            <div className="relative">
-                              <input
-                                type="text"
-                                value={
-                                  availability &&
-                                  availability.find((elem) => Object.keys(elem)[0] === Object.keys(eachObj)[0]).time[0].end === ""
-                                    ? "17:00"
-                                    : startEndObj.end
-                                }
-                                id={`${Object.keys(eachObj)[0]}+end+${timeIndex}`}
-                                onClick={(e) => displayTimeDropdown(e.target.id, e.currentTarget)}
-                                onChange={(e) => e.target.value}
-                                className="border border-gray-700 rounded-lg w-16 h-7 text-center"
-                              />
-                              {selectedItem === `${Object.keys(eachObj)[0]}+end+${timeIndex}` && (
-                                <TimeDropdown
-                                  selectedItem={selectedItem}
-                                  timeIndex={timeIndex}
-                                  from={"weekly"}
-                                />
-                              )}
-                            </div>
-                            <div className="flex justify-center items-center w-7 h-7 hover:bg-gray-200 hover:rounded-full">
-                              <svg onClick={(e) => handleMethod(e, "delete", Object.keys(eachObj)[0], startEndObj)} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-5 h-5">
-                                <path strokeLinecap="round" strokeLinejoin="round" d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"/>
-                              </svg>
-                            </div>
-                          </div>
-                        );
-                      })
-                  ) : (
-                    <div className="text-gray-400 w-[80%] ml-5 h-7">Unavailable</div>
-                  )}
+                          );
+                        })
+                    ) : (
+                      <div className="text-gray-400 w-[80%] ml-5 h-7">Unavailable</div>
+                    )}
                 </div>
               </div>
